Blam cleanup part 1
