+++ xlive/Blam/Cache/DataTypes.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "DataTypes\DatumIndex.h"
+#include "DataTypes\Reflexive.h"
+#include "DataTypes\Macros.h"
+#include "DataTypes\StringID.h"
+#include "DataTypes\Strings.h"
+#include "DataTypes\TagRef.h"
+#include "DataTypes\DataRef.h"
+#include "DataTypes\BlamTag.h"
+#include "DataTypes\TagBlock.h"
