--- xlive/Blam/Cache/DataTypes/DataTypes.h
@@ -1,9 +1,0 @@
-#pragma once
-#include "DatumIndex.h"
-#include "Reflexive.h"
-#include "Macros.h"
-#include "StringID.h"
-#include "Strings.h"
-#include "TagRef.h"
-#include "DataRef.h"
-#include "BlamTag.h"
