+++ xlive/Blam/Cache/TagGroups.hpp
@@ -1,9 +1,5 @@
 #pragma once
-#ifndef TAG_GROUPS_H
-#define TAG_GROUPS_H
 #include "Blam\Enums\Tags\TagGroups.h"
-#include "Blam\Cache\DataTypes\Strings.h"
-
 
 //Base definition for a TagGroup
 template <const LONG GroupTag>
@@ -14,8 +10,4 @@ struct TagGroup
 
 	//TagGroup Full Name
 	static std::string LongName; //Need a quite way to define for individual tags
-
 };
-
-
-#endif
