+++ xlive/Blam/Cache/TagGroups/ai_mission_dialogue_definition.hpp
@@ -1,13 +1,15 @@
 #pragma once
-#pragma pack(push,1)
 #include "..\Blam\Cache\DataTypes\DataTypes.h"
 #include "..\Blam\Cache\TagGroups.hpp"
 #include "..\Blam\Cache\DataTypes\TagBlock.h"
+
 /*********************************************************************
 * name: ai_mission_dialogue
 * group_tag : mdlg
 * header size : 8
 * *********************************************************************/
+
+#pragma pack(push,1)
 struct s_ai_mission_dialogue_group_definition :TagGroup<'mdlg'>
 {
 	struct s_lines_block
