+++ xlive/Blam/Cache/TagGroups/crate_definition.hpp
@@ -1,11 +1,13 @@
 #pragma once
-#pragma pack(push,1)
 #include "object_definition.hpp"
+
 /*********************************************************************
 * name: crate
 * group_tag : bloc
 * header size : 192
 * *********************************************************************/
+
+#pragma pack(push,1)
 struct s_crate_group_definition :TagGroup<'bloc'>
 {
 	s_object_group_definition objectTag;
@@ -17,5 +19,4 @@ struct s_crate_group_definition :TagGroup<'bloc'>
 	PAD(0x2);//0xBE
 };
 TAG_GROUP_SIZE_ASSERT(s_crate_group_definition, 0xC0);
-
 #pragma pack(pop)
