+++ xlive/Blam/Cache/TagGroups/device_light_fixture_definition.hpp
@@ -1,11 +1,13 @@
 #pragma once
-#pragma pack(push,1)
 #include "device_definition.hpp"
+
 /*********************************************************************
 * name: device_light_fixture
 * group_tag : lifi
 * header size : 284
 * *********************************************************************/
+
+#pragma pack(push,1)
 struct s_device_light_fixture_group_definition :TagGroup<'lifi'>
 {
 	s_device_group_definition deviceTag;
