+++ xlive/Blam/Cache/Tags/tag_base.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include"..\DataTypes\DataTypes.h"
+#include"..\DataTypes.h"
 #include"..\TagGroups.hpp"
 #include"..\Blam\Maths\Maths.h"
